Test Pulley on CI on 32-bit architectures
This commit extends our CI for i686 and armv7 to test the Pulley
backend, namely the full `*.wast` test suite as well as the `wasmtime`
crate itself. Note that many `*.wast` tests are still expected to fail
at this time.

This involved fixing a number of 32-vs-64 bit issues throughout the test
suite in various location in this commit.

+++ cranelift/jit/src/compiled_blob.rs
@@ -102,7 +102,7 @@ impl CompiledBlob {
                     let at_page = (at as usize) & !0xfff;
                     let pcrel = (what_page as isize).checked_sub(at_page as isize).unwrap();
                     assert!(
-                        (-1 << 32) <= pcrel && pcrel < (1 << 32),
+                        (-1 << 32) <= (pcrel as i64) && (pcrel as i64) < (1 << 32),
                         "can't reach GOT page with ±4GB `adrp` instruction"
                     );
                     let val = pcrel >> 12;
